moved filtering&sorting logic
moved the filtering and sorting logic to the page ProdactListPage
